requirements.txt
fastapi
uvicorn
scikit-learn
pandas
joblib
requests
python-multipart
pypdf
Pillow